PMM-7873 move common messages
