--- public/app/percona/backup/components/BackupInventory/BackupInventory.messages.ts
@@ -1,26 +1,0 @@
-export const Messages = {
-  noData: 'No Backups found',
-  columns: {
-    name: 'Backup name',
-    created: 'Created',
-    location: 'Location',
-    vendor: 'Vendor',
-    status: 'Status',
-    actions: 'Actions',
-  },
-  status: {
-    invalid: 'Invalid',
-    pending: 'Pending',
-    inProgress: 'In progress',
-    paused: 'Paused',
-    success: 'Success',
-    error: 'Error',
-  },
-  dataModel: {
-    invalid: 'Invalid',
-    physical: 'Physical',
-    logical: 'Logical',
-  },
-  actions: 'Actions',
-  add: 'Add',
-};
