+++ public/app/percona/backup/components/BackupInventory/BackupInventory.tsx
@@ -11,13 +11,10 @@ import { AddBackupFormProps } from './AddBackupModal/AddBackupModal.types';
 import { Status } from './Status';
 import { BackupInventoryActions } from './BackupInventoryActions';
 import { BackupCreation } from './BackupCreation';
-import { Messages } from './BackupInventory.messages';
+import { Messages } from '../../Backup.messages';
 import { Backup } from './BackupInventory.types';
 import { BackupInventoryService } from './BackupInventory.service';
 import { getStyles } from './BackupInventory.styles';
-
-const { columns, noData } = Messages;
-const { name, created, location, vendor, status, actions } = columns;
 
 export const BackupInventory: FC = () => {
   const [pending, setPending] = useState(false);
@@ -27,33 +24,33 @@ export const BackupInventory: FC = () => {
   const columns = useMemo(
     (): Column[] => [
       {
-        Header: name,
+        Header: Messages.backupInventory.table.columns.name,
         accessor: 'name',
         id: 'name',
         width: '250px',
         Cell: ({ row, value }) => <ExpandableCell row={row} value={value} />,
       },
       {
-        Header: vendor,
+        Header: Messages.backupInventory.table.columns.vendor,
         accessor: ({ vendor }: Backup) => DATABASE_LABELS[vendor],
         width: '150px',
       },
       {
-        Header: created,
+        Header: Messages.backupInventory.table.columns.created,
         accessor: 'created',
         Cell: ({ value }) => <BackupCreation date={value} />,
       },
       {
-        Header: location,
+        Header: Messages.backupInventory.table.columns.location,
         accessor: 'locationName',
       },
       {
-        Header: status,
+        Header: Messages.backupInventory.table.columns.status,
         accessor: 'status',
         Cell: ({ value }) => <Status status={value} />,
       },
       {
-        Header: actions,
+        Header: Messages.backupInventory.table.columns.actions,
         accessor: 'id',
         Cell: ({ row }) => <BackupInventoryActions onBackup={onBackupClick} backup={row.original as Backup} />,
         width: '80px',
@@ -129,7 +126,7 @@ export const BackupInventory: FC = () => {
         data={data}
         totalItems={data.length}
         columns={columns}
-        emptyMessage={noData}
+        emptyMessage={Messages.backupInventory.table.noData}
         pendingRequest={pending}
         renderExpandedRow={renderSelectedSubRow}
       ></Table>
