--- public/app/percona/backup/components/RestoreHistory/RestoreHistory.messages.ts
@@ -1,6 +1,0 @@
-export const Messages = {
-  noData: 'No Restores found',
-  columns: {
-    started: 'Started At',
-  },
-};
